# salesforce-deployer/src/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: your.email@example.com
  description: |-
    API for extracting, managing, and deploying Salesforce metadata schemas.
    This API enables seamless management of Salesforce configurations across environments.
  license:
    name: MIT
  title: Salesforce Schema Deployer API
  version: 0.1.0
servers:
- description: Development server
  url: http://localhost:8000/api/v1
- description: Mock server for testing
  url: http://localhost:8001/api/v1
...
paths:
...
  /schemas/extract:
    post:
      description: "Extracts metadata schema components from a specified Salesforce\
        \ organization, with options to filter by component types and include standard\
        \ or managed components."
      operationId: extractSchema
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaExtractionRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaExtractionResponse'
          description: Schema extraction successful
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Extract schema from a Salesforce organization
      tags:
      - schemas
  /schemas/{schemaId}:
    delete:
      description: Permanently removes a schema definition and all its associated
        metadata components from the system.
      operationId: deleteSchema
      parameters:
      - explode: false
        in: path
        name: schemaId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Schema deleted successfully
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Schema not found
      security:
      - BearerAuth: []
      summary: Delete schema by ID
      tags:
      - schemas
    get:
      description: "Retrieves the complete metadata schema definition by its unique\
        \ identifier, including all component definitions and relationships."
      operationId: getSchema
      parameters:
      - explode: false
        in: path
        name: schemaId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schema'
          description: Schema found
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Schema not found
      security:
      - BearerAuth: []
      summary: Get schema by ID
      tags:
      - schemas
    put:
      description: "Updates an existing schema definition with new metadata components,\
        \ properties, or relationships while preserving its identifier."
      operationId: updateSchema
      parameters:
      - explode: false
        in: path
        name: schemaId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schema'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schema'
          description: Schema updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Schema not found
      security:
      - BearerAuth: []
      summary: Update schema by ID
      tags:
      - schemas
  /schemas/compare:
    post:
      description: "Performs a detailed comparison between two schemas, identifying\
        \ added, modified, and removed components with field-level change tracking."
      operationId: compareSchemas
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaComparisonRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaComparisonResponse'
          description: Schema comparison completed
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request
      security:
      - BearerAuth: []
      summary: Compare two schemas
      tags:
      - schemas
  /schemas/validate:
    post:
      description: "Analyzes a schema for validity, checking component dependencies,\
        \ references, and compatibility with target Salesforce organizations."
      operationId: validateSchema
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaValidationRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaValidationResponse'
          description: Schema validation completed
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request
      security:
      - BearerAuth: []
      summary: Validate a schema
      tags:
      - validation
...
  /organizations:
    get:
      description: Retrieves all Salesforce organizations that have been configured
        for schema extraction and deployment operations.
      operationId: getOrganizations
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Organization'
                type: array
          description: List of organizations retrieved successfully
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Get a list of connected Salesforce organizations
      tags:
      - organizations
    post:
      description: "Registers a new Salesforce organization with the system, including\
        \ connection credentials and organization metadata."
      operationId: addOrganization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: Organization added successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request
      security:
      - BearerAuth: []
      summary: Add a new Salesforce organization
      tags:
      - organizations
  /organizations/{orgId}:
    delete:
      description: "Removes a Salesforce organization from the system, including all\
        \ stored credentials and connection information."
      operationId: deleteOrganization
      parameters:
      - explode: false
        in: path
        name: orgId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Organization deleted successfully
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Organization not found
      security:
      - BearerAuth: []
      summary: Delete organization
      tags:
      - organizations
    get:
      description: "Retrieves detailed information about a specific Salesforce organization,\
        \ including its connection status and metadata."
      operationId: getOrganizationById
      parameters:
      - explode: false
        in: path
        name: orgId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: Organization details retrieved successfully
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Organization not found
      security:
      - BearerAuth: []
      summary: Get organization details by ID
      tags:
      - organizations
    put:
      description: "Updates the configuration, credentials, or metadata for an existing\
        \ Salesforce organization connection."
      operationId: updateOrganization
      parameters:
      - explode: false
        in: path
        name: orgId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: Organization updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Organization not found
      security:
      - BearerAuth: []
      summary: Update organization details
      tags:
      - organizations
components:
  schemas:
...
    ErrorResponse:
      example:
        details:
        - field: username
          message: Field is required
        - field: username
          message: Field is required
        error: INVALID_SESSION_ID
        message: Session expired or invalid
      properties:
        error:
          description: Error code
          example: INVALID_SESSION_ID
          title: error
          type: string
        message:
          description: Error message
          example: Session expired or invalid
          title: message
          type: string
        details:
          description: Detailed error information
          items:
            $ref: '#/components/schemas/ErrorResponse_details_inner'
          title: details
          type: array
      required:
      - error
      - message
      title: ErrorResponse
      type: object
    SchemaExtractionRequest:
      example:
        includeStandard: false
        includeManaged: false
        filter: "CustomObject:Account,Contact"
        components:
        - objects
        - objects
        orgId: 123e4567-e89b-12d3-a456-426614174000
      properties:
        orgId:
          description: ID of the organization to extract schema from
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          title: orgId
          type: string
        components:
          description: "List of components to extract. If not provided, all components\
            \ will be extracted."
          items:
            enum:
            - objects
            - fields
            - workflows
            - dashboards
            - flows
            example: objects
            type: string
          title: components
          type: array
        includeStandard:
          default: false
          description: Whether to include standard objects
          title: includeStandard
          type: boolean
        includeManaged:
          default: false
          description: Whether to include managed package components
          title: includeManaged
          type: boolean
        filter:
          description: Filter expression to limit extracted components
          example: "CustomObject:Account,Contact"
          title: filter
          type: string
      required:
      - orgId
      title: SchemaExtractionRequest
      type: object
    SchemaExtractionResponse:
      example:
        stats:
          flows: 2
          objects: 10
          workflows: 5
          fields: 50
          dashboards: 3
        schemaId: 123e4567-e89b-12d3-a456-426614174000
        progress: 100
        message: Schema extraction completed successfully
        status: completed
      properties:
        schemaId:
          description: ID of the extracted schema
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          title: schemaId
          type: string
        status:
          description: Status of the extraction process
          enum:
          - completed
          - in_progress
          - failed
          example: completed
          title: status
          type: string
        progress:
          description: Progress percentage
          example: 100
          maximum: 100
          minimum: 0
          title: progress
          type: integer
        message:
          description: Additional information
          example: Schema extraction completed successfully
          title: message
          type: string
        stats:
          $ref: '#/components/schemas/SchemaExtractionResponse_stats'
      required:
      - schemaId
      - status
      title: SchemaExtractionResponse
      type: object
    Schema:
      example:
        createdAt: 2025-04-21T14:30:00Z
        components:
          workflowRules:
            key:
              criteria: Account.Status__c = 'Active'
              name: Account_Status_Update
              active: true
              description: Updates Account status based on criteria
              actions:
              - field: Status__c
                name: UpdateAccountStatus
                type: FieldUpdate
                value: Processed
              - field: Status__c
                name: UpdateAccountStatus
                type: FieldUpdate
                value: Processed
              object: Account
          flows:
            key:
              name: Lead_Process
              description: Automates lead qualification
              label: Lead Qualification Process
              processType: AutoLaunchedFlow
              triggerType: ""
              eventType: ""
              status: Active
          objects:
            key:
              custom: false
              name: Account
              description: Standard Account object
              nameField: Name
              label: Account
              fields:
                key:
                  relationshipName: ""
                  defaultValue: Default value
                  unique: false
                  name: CustomField__c
                  length: 255
                  externalId: false
                  description: A custom text field
                  formula: ""
                  label: Custom Field
                  type: Text
                  required: false
                  referenceTo: ""
              sharingModel: ReadWrite
          dashboards:
            key:
              components:
              - reportName: SalesReport
                name: SalesChart
                type: Chart
              - reportName: SalesReport
                name: SalesChart
                type: Chart
              name: Sales_Dashboard
              description: Key sales metrics dashboard
              label: Sales Performance
              folderName: Sales Dashboards
        apiVersion: "56.0"
        sourceOrgId: 123e4567-e89b-12d3-a456-426614174000
        name: Production Schema April 2025
        description: Current production schema as of April 2025
        id: 123e4567-e89b-12d3-a456-426614174000
        updatedAt: 2025-04-21T14:30:00Z
      properties:
        id:
          description: Schema ID
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          title: id
          type: string
        name:
          description: Schema name
          example: Production Schema April 2025
          title: name
          type: string
        description:
          description: Schema description
          example: Current production schema as of April 2025
          title: description
          type: string
        createdAt:
          description: Schema creation timestamp
          example: 2025-04-21T14:30:00Z
          format: date-time
          title: createdAt
          type: string
        updatedAt:
          description: Schema last update timestamp
          example: 2025-04-21T14:30:00Z
          format: date-time
          title: updatedAt
          type: string
        sourceOrgId:
          description: ID of the source organization
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          title: sourceOrgId
          type: string
        apiVersion:
          description: Salesforce API version
          example: "56.0"
          title: apiVersion
          type: string
        components:
          $ref: '#/components/schemas/Schema_components'
      required:
      - id
      - name
      title: Schema
      type: object
    SchemaComparisonRequest:
      example:
        components:
        - objects
        - objects
        sourceSchemaId: 123e4567-e89b-12d3-a456-426614174000
        targetSchemaId: 223e4567-e89b-12d3-a456-426614174000
      properties:
        sourceSchemaId:
          description: ID of the source schema
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          title: sourceSchemaId
          type: string
        targetSchemaId:
          description: ID of the target schema
          example: 223e4567-e89b-12d3-a456-426614174000
          format: uuid
          title: targetSchemaId
          type: string
        components:
          description: "List of components to compare. If not provided, all components\
            \ will be compared."
          items:
            enum:
            - objects
            - fields
            - workflows
            - dashboards
            - flows
            example: objects
            type: string
          title: components
          type: array
      required:
      - sourceSchemaId
      - targetSchemaId
      title: SchemaComparisonRequest
      type: object
    SchemaComparisonResponse:
      example:
        differences:
        - componentType: field
          changeType: modified
          targetValue: "{}"
          targetPath: objects.Account.fields.CustomField__c
          componentName: Account.CustomField__c
          sourceValue: "{}"
          sourcePath: objects.Account.fields.CustomField__c
          fieldChanges:
          - field: length
            changeType: modified
            targetValue:
              length: 255
            sourceValue:
              length: 80
          - field: length
            changeType: modified
            targetValue:
              length: 255
            sourceValue:
              length: 80
        - componentType: field
          changeType: modified
          targetValue: "{}"
          targetPath: objects.Account.fields.CustomField__c
          componentName: Account.CustomField__c
          sourceValue: "{}"
          sourcePath: objects.Account.fields.CustomField__c
          fieldChanges:
          - field: length
            changeType: modified
            targetValue:
              length: 255
            sourceValue:
              length: 80
          - field: length
            changeType: modified
            targetValue:
              length: 255
            sourceValue:
              length: 80
      properties:
        differences:
          items:
            $ref: '#/components/schemas/SchemaComparisonResponse_differences_inner'
          title: differences
          type: array
      required:
      - differences
      title: SchemaComparisonResponse
      type: object
...
    SchemaValidationResponse:
      example:
        valid: true
        dependencyGraph:
          key:
          - dependencyGraph
          - dependencyGraph
        warnings:
        - severity: error
          componentType: field
          path: objects.Account.fields.CustomField__c
          componentName: Account.CustomField__c
          message: Field 'CustomField__c' references non-existent object 'NonExistentObject__c'
        - severity: error
          componentType: field
          path: objects.Account.fields.CustomField__c
          componentName: Account.CustomField__c
          message: Field 'CustomField__c' references non-existent object 'NonExistentObject__c'
        errors:
        - severity: error
          componentType: field
          path: objects.Account.fields.CustomField__c
          componentName: Account.CustomField__c
          message: Field 'CustomField__c' references non-existent object 'NonExistentObject__c'
        - severity: error
          componentType: field
          path: objects.Account.fields.CustomField__c
          componentName: Account.CustomField__c
          message: Field 'CustomField__c' references non-existent object 'NonExistentObject__c'
      properties:
        valid:
          description: Whether the schema is valid
          example: true
          title: valid
          type: boolean
        errors:
          description: Validation errors (if any)
          items:
            $ref: '#/components/schemas/ValidationError'
          title: errors
          type: array
        warnings:
          description: Validation warnings
          items:
            $ref: '#/components/schemas/ValidationError'
          title: warnings
          type: array
        dependencyGraph:
          additionalProperties:
            items:
              type: string
            type: array
          description: Graph of component dependencies
          title: dependencyGraph
          type: object
      required:
      - valid
      title: SchemaValidationResponse
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http